Cache removed and slipts classes
