+++ SAPSec.Web/Filters/ExpectedSchoolPhase.cs
@@ -1,0 +1,7 @@
+namespace SAPSec.Web.Filters;
+
+public enum ExpectedSchoolPhase
+{
+    Primary,
+    Secondary
+}
