+++ SAPSec.Web/Services/RequestSchoolAccessor.cs
@@ -1,6 +1,5 @@
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
-using SAPSec.Web.Filters;
 
 namespace SAPSec.Web.Services;
 
@@ -8,13 +7,6 @@ public class RequestSchoolAccessor(ISchoolDetailsService schoolDetailsService) :
 {
     public async Task<SchoolDetails> GetAsync(HttpContext? httpContext, string urn)
     {
-        if (ValidatedSchoolDetailsStore.TryGet(httpContext, urn, out var school))
-        {
-            return school;
-        }
-
-        school = await schoolDetailsService.GetByUrnAsync(urn);
-        ValidatedSchoolDetailsStore.Set(httpContext, urn, school);
-        return school;
+        return await schoolDetailsService.GetByUrnAsync(urn);
     }
 }
